test: 검색 test가 제대로 되지 않던 오류 수정 및 test 코드 분리
